refactor: exception 구조 변경
