+++ meongcare/src/main/java/com/meongcare/common/error/exception/clientError/EntityNotFoundException.java
@@ -1,8 +1,8 @@
-package com.meongcare.common.error.exception;
+package com.meongcare.common.error.exception.clientError;
 
 import com.meongcare.common.error.ErrorCode;
 
-public class EntityNotFoundException extends BaseException{
+public class EntityNotFoundException extends BadRequestException {
     public EntityNotFoundException(ErrorCode errorCode) {
         super(errorCode);
     }
