+++ meongcare/src/main/java/com/meongcare/domain/dog/domain/DogRepository.java
@@ -1,7 +1,7 @@
 package com.meongcare.domain.dog.domain;
 
 import com.meongcare.common.error.ErrorCode;
-import com.meongcare.common.error.exception.EntityNotFoundException;
+import com.meongcare.common.error.exception.clientError.EntityNotFoundException;
 import com.meongcare.domain.dog.domain.entity.Dog;
 import com.meongcare.domain.member.domain.entity.Member;
 import org.springframework.data.jpa.repository.JpaRepository;
